+++ src/utils/index.ts
@@ -16,3 +16,4 @@ export * from "./NumberUtils";
 export * from "./DeploymentUtils";
 export * from "./FormattingUtils";
 export * from "./BlockExplorerUtils";
+export * from "./BigNumberUtils";
